# cloud/mission_svc/openapi.json
{
  "components": {
    "schemas": {
      "AuditEntry": {
        "description": "審計軌跡一筆(供 GET /api/v1/audit,admin 稽核檢視)。",
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "resource_type": {
            "title": "Resource Type",
            "type": "string"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "source_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ip"
          }
        },
        "required": [
          "id",
          "at",
          "actor",
          "action",
          "resource_type"
        ],
        "title": "AuditEntry",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Mission": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_item": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Item"
          },
          "dispatched_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispatched At"
          },
          "drone_id": {
            "title": "Drone Id",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mission_id": {
            "title": "Mission Id",
            "type": "string"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "route_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Id"
          },
          "rtl_after_last": {
            "title": "Rtl After Last",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/MissionStatus"
          },
          "total_items": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Items"
          },
          "waypoints": {
            "items": {
              "$ref": "#/components/schemas/Waypoint"
            },
            "title": "Waypoints",
            "type": "array"
          }
        },
        "required": [
          "id",
          "mission_id",
          "org_id",
          "drone_id",
          "status",
          "waypoints",
          "rtl_after_last",
          "created_at"
        ],
        "title": "Mission",
        "type": "object"
      },
      "MissionCommandRequest": {
        "properties": {
          "command": {
            "$ref": "#/components/schemas/CommandKind"
          }
        },
        "required": [
          "command"
        ],
        "title": "MissionCommandRequest",
        "type": "object"
      },
      "MissionCreate": {
        "description": "由 route + 目標機建立任務(凍結 route 當下的航點)。",
        "properties": {
          "drone_id": {
            "title": "Drone Id",
            "type": "string"
          },
          "route_id": {
            "format": "uuid",
            "title": "Route Id",
            "type": "string"
          }
        },
        "required": [
          "route_id",
          "drone_id"
        ],
        "title": "MissionCreate",
        "type": "object"
      },
      "MissionStatus": {
        "enum": [
          "created",
          "dispatched",
          "received",
          "uploaded",
          "in_progress",
          "paused",
          "completed",
          "failed"
        ],
        "title": "MissionStatus",
        "type": "string"
      },
      "Route": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "rtl_after_last": {
            "title": "Rtl After Last",
            "type": "boolean"
          },
          "waypoints": {
            "items": {
              "$ref": "#/components/schemas/Waypoint"
            },
            "title": "Waypoints",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "org_id",
          "waypoints",
          "rtl_after_last",
          "created_at"
        ],
        "title": "Route",
        "type": "object"
      },
      "RouteCreate": {
        "description": "建立航線。org_id 不在此——租戶由呼叫者 JWT claim 決定(不採信 client 傳入)。",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rtl_after_last": {
            "default": true,
            "title": "Rtl After Last",
            "type": "boolean"
          },
          "waypoints": {
            "items": {
              "$ref": "#/components/schemas/Waypoint"
            },
            "minItems": 1,
            "title": "Waypoints",
            "type": "array"
          }
        },
        "required": [
          "name",
          "waypoints"
        ],
        "title": "RouteCreate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "mission-svc",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/audit": {
      "get": {
        "operationId": "list_audit_api_v1_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "resource_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEntry"
                  },
                  "title": "Response List Audit Api V1 Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Audit"
      }
    },
    "/api/v1/missions": {
      "get": {
        "operationId": "list_missions_api_v1_missions_get",
        "parameters": [
          {
            "in": "query",
            "name": "drone_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Drone Id"
            }
          },
          {
            "description": "僅 admin:限定單一租戶(略則看全部)",
            "in": "query",
            "name": "org",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "僅 admin:限定單一租戶(略則看全部)",
              "title": "Org"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Mission"
                  },
                  "title": "Response List Missions Api V1 Missions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Missions"
      },
      "post": {
        "operationId": "create_mission_api_v1_missions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mission"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Mission"
      }
    },
    "/api/v1/missions/{mission_pk}": {
      "get": {
        "operationId": "get_mission_api_v1_missions__mission_pk__get",
        "parameters": [
          {
            "in": "path",
            "name": "mission_pk",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Mission Pk",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mission"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Mission"
      }
    },
    "/api/v1/missions/{mission_pk}/command": {
      "post": {
        "operationId": "command_mission_api_v1_missions__mission_pk__command_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_pk",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Mission Pk",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mission"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Command Mission"
      }
    },
    "/api/v1/missions/{mission_pk}/dispatch": {
      "post": {
        "operationId": "dispatch_mission_api_v1_missions__mission_pk__dispatch_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_pk",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Mission Pk",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mission"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Dispatch Mission"
      }
    },
    "/api/v1/routes": {
      "get": {
        "operationId": "list_routes_api_v1_routes_get",
        "parameters": [
          {
            "description": "僅 admin:限定單一租戶(略則看全部)",
            "in": "query",
            "name": "org",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "僅 admin:限定單一租戶(略則看全部)",
              "title": "Org"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Route"
                  },
                  "title": "Response List Routes Api V1 Routes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Routes"
      },
      "post": {
        "operationId": "create_route_api_v1_routes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Route"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Route"
      }
    },
    "/api/v1/routes/{route_id}": {
      "get": {
        "operationId": "get_route_api_v1_routes__route_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Route"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Route"
      }
    },
    "/api/v1/usage": {
      "get": {
        "operationId": "get_usage_api_v1_usage_get",
        "parameters": [
          {
            "description": "僅 admin:查指定租戶(略則查本 org)",
            "in": "query",
            "name": "org",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "僅 admin:查指定租戶(略則查本 org)",
              "title": "Org"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Usage"
      }
    },
...
  }
}